# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory-API",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://jeremybecker.ch/",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/location": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "List locations",
                "description": "Get a List of all locations.",
                "operationId": "listLocation",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Create location",
                "description": "Create a new location.",
                "operationId": "storeLocation",
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "description": "The branch of the company where the location is.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "room",
                        "in": "path",
                        "description": "The room the location is.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "shelf",
                        "in": "path",
                        "description": "The shelf the location is in.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "compartment",
                        "in": "path",
                        "description": "The compartment in the shelf of the location.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/location/{id}": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Show location",
                "description": "Display a location",
                "operationId": "showLocation",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Location"
                ],
                "summary": "Delete location",
                "description": "Delete a location.",
                "operationId": "deleteLocation",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Location"
                ],
                "summary": "Update location",
                "description": "Update a new location.",
                "operationId": "updateLocation",
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "description": "The branch of the company where the location is.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "room",
                        "in": "path",
                        "description": "The room the location is.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "shelf",
                        "in": "path",
                        "description": "The shelf the location is in.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "compartment",
                        "in": "path",
                        "description": "The compartment in the shelf of the location.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/tag": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "List tags",
                "description": "Get a List of all tags",
                "operationId": "listTag",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "tags": [
                    "Tag"
                ],
                "summary": "Create tag",
                "description": "Create a new Tag",
                "operationId": "storeTag",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The display name of the tag.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "inventory_item_id",
                        "in": "path",
                        "description": "The inventory item the tag is associated with.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/tag/{id}": {
            "get": {
                "tags": [
                    "Tag"
                ],
                "summary": "Show tag",
                "description": "Display a single tag",
                "operationId": "showTag",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tag"
                ],
                "summary": "Delete tag",
                "description": "Delete a tag.",
                "operationId": "deleteTag",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tag"
                ],
                "summary": "Update tag",
                "description": "Update variables of a tag.",
                "operationId": "updateTag",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "The display name of the tag.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "inventory_item_id",
                        "in": "path",
                        "description": "The inventory item the tag is associated with.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    }
}